% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.sum.R
\name{class.sum}
\alias{class.sum}
\title{Class Sum Function}
\usage{
class.sum(truth, predicted)
}
\description{
none
}
\examples{
class.sum()
}
\author{
Richard Cutler, \email{richard.cutler@usu.edu}
}
\keyword{logistic}
